+++ tests/workflows/basic_parent_context/tests/test_basic_wrapper.py
@@ -1,7 +1,7 @@
 from uuid import uuid4
 
-from vellum.core.context import get_execution_context, wrapper_execution_parent_context
-from vellum.workflows.events.types import NodeParentContext, ParentContext
+from vellum.workflows.context import get_execution_context, wrapper_execution_parent_context
+from vellum.workflows.events.types import NodeParentContext
 
 from tests.workflows.basic_parent_context.basic_workflow import StartNode
